\name{mvctm-package}
\alias{mvctm-package}
\docType{package}
\title{
Multivariate Variance Components Tests for Multilevel Data
}
\description{
Permutation tests for variance components for 2-level, 3-level and 4-level data with univariate or multivariate responses.
}
\details{
\tabular{ll}{
Package: \tab mvctm\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2017-03-08\cr
}
}
\author{
Denis Larocque

Maintainer: Denis Larocque \code{<denis.larocque@hec.ca>}
}

\references{
Larocque, D., Nevalainen, J. and Oja, H. (2017). Multivariate Variance Components Tests for Multilevel Data: The mvctm package. Les Cahiers du GERAD G-2017-16.
}



